requirements.txt
streamlit
transformers
torch
pdfplumber
python-pptx
pandas